Implement error dialog, display errors better
 - Add dialog box fragment to display errors
 - Remove unused Loader implementation
 - Display username/password errors below field

# Explanation
There are three types of errors in the login field:
 - Username validation/authentication errors
 - Password validation/authentication errors
 - Login/connection errors

Validation and authentication errors are specific to the data being entered, and should display below the text field in the error label area.

Login/connection errors are more general, and interrupt the flow of the application. They should also have a 'try again' option to resolve the issue. As such, we now have an error dialog that has a 'cancel' and 'try again' button.

# Issues Encountered
Displaying a dialog box on Android is a little trickier than you'd think. This is because these boxes have to scale, respond to orientation changes, and respond to lots of different screen events. As such, they're handled by Fragments, which are like mini-activities.

There's an easy way to just build a dialog box, but I'm trying to avoid shortcuts for the sake of having a robust, well-made application (and thus avoid the UX issues which plague the existing solutions). So I had to implement the DialogFragment class.

The first issue I had was removing the title from the box. Because I appropriated my code from StackOverflow, my box had a title I didn't need. After spending an hour or so looking around for how to do this, I realised that simply not passing a title element to the box in construction removed it.

The second issue I had was sending the results back to the Activity that called it. The first solution from StackOverflow only worked when a Fragment was calling another Fragment. Eventually, after a lot of trial and error, I managed to wrap my head around the Listener lifecycle and how to send these events between classes. As a result, my ErrorDialog sends the results back to _whatever_ called it, so long as it implements the onErrorDialogDismiss listener. If it doesn't, then we just dismiss the box normally.