Improve sign in button
 - Colour in the background of the sign in button
 - Make the button text white
 - Make it wrap the content, and push it to the side
 - Fix the shadow clipping by removing useless container views

# Issues Encountered
The first big problem was the button's shadow was clipping off on android 5.0 and above. After a struggle, the solution to this was to move everything into RelativeLayouts, remove the useless ScrollView, and apply `android:clipToPadding="false"` to the parent layout.

This way we keep the padding on the view, but we get rid of the annoying clipping.

The second major issue was the text colour. On pre-lollipop devices the text colour is black for the AppCompat Button, but on lollipop and above it's white (as expected). To solve this, I created a style theme for the button which colours the background in purple and the foreground in white, and applied it to both configurations. This overrode the default black on pre-lollipop, and maintained the white on lollipop and above.