Fix progress spinner
# Explanation
The progress spinner has been in this activity since I started it with the default template. However, it's kinda been in the background and kinda broken.

Good UX requires that, when the app is doing something interesting, we show the progress. We also can't just switch to the next view, in case there's a connection issue or a wrong password. So the spinner helps the user understand that something's going on in the background.

Also, it was important to have a short fade time, so that the user understands that their action is being processed and reacted to. Android has safe constants for this which were easily to deal with.

 - Center the progress spinner
 - Correctly fade between views
 - Hide keyboard when spinner appears