Add credentials to AccountManager
 - Implement AccountAuthenticator for Omni Sync Server
 - Include account service in Settings > Accounts
 - Simplified logging format
 - Fix @NonNull warning for ErrorDialog

# Explanation
Best practice on Android for storing an account that needs to sync is to use the AccountManager service to register and provide account services. The advantage here is that the credentials are 100% inaccessible from apps that don't have permission (unless you're running with root - but you'd have bigger problems then anyway).

The issue with the Omni Sync Server is that we have to store the credentials in plaintext in lieu of the sync server implementing authentication tokens. But we have no other option, so storing them this way is the safest way to do so.

Oh, and I think this is finally the end of the login activity! That's an awesome little milestone to have hit this early into development!

# Issues Encountered
The big issue I had was dealing with implementation - it's all a little confusing for a first-time Android developer, especially because I don't fully understand how the Intent system works. The other issue with implementation was adapting existing implementations to deal with using plaintext credentials instead of an auth token.