Totally restructured code
 - Moved classes into new packages
 - Moved some AsyncTasks out of their activities and into new classes
 - Added universally accessible account type string

 # Explanation
 My code was getting a little rusty through the project, and I was starting to make a lot of classes. I knew I'd have to start organising the files at some point, and it made sense after making my fourth account-related class that it'd be a good idea to group everything.

 The code in OmniSyncLogin in particular relies instead on callback functions (known as listeners in Java) to deal with returned data. This is the pattern I'll be implementing from here on in throughout my project.

# Issues Encountered
The big issue here was just keeping track of everything. It's easy enough to manipulate some code within one class for a while, there's not much mental strain in that. But dealing with a serious restructure requires you to have a deep and intense knowledge of your code, AND its structure as well. So that was a bit of a challenge, but you come out of it with a better knowledge of how everything works.

Speaking of, I'm really getting the hang of Java/Android development now. It's a little tricky to pick up, but once you understand what the best practices are and how to use them, it's really rewarding. The app I eventually build is going to be really really cool but also really robust now that I'm doing this stuff.