Add file list downloader
# Explanation
GetFilesToDownload is an AsyncTask which heads to the Omni Sync Server and gets a list of all files under /OmniFocus.ofocus/, which are to be downloaded by our client.

It also sifts through the file list and determines which files it actually needs to download. Given the uncontrollable state of the sync format, this is a little tricky, and this method will need to be extended in future. But for our purposes right now, it's pretty okay.