Add app icon
# Explanation
I was so sick of the app just sitting there with the annoying default icon (I was going to write 'awful', but it's actually pretty okay). Because the icon shows up in a few places now (launcher, accounts), I figured I needed to change it.

For more details see the theory document.