Improve error login using Log.x() third parameter
Log.x() (where x = e, i, etc.) accepts a third parameter, a throwable. This is so that Log.x() can handle errors and your own text in one line, and give you an idea of where the error is coming from with your own tag. This commit updates the error logging to support this.