Add basic data model
# Explanation
To read data into our application, each class of item in the OmniFocus world needs a corresponding class to have its properties read into in Android. This way we can convert easily between data being in XML, UI and SQL.

Building this was kind of easy, but I'm not sure that I'll stick with the decisions I made. Effectively, I intend for one of these classes to be made for each thing that's parsed by the ContentsConverter. I chose the variables in their constructors to be the variables that are found within the sync data XMLs, but converted to Java-native types.

So the model for all of this is that the XML is parsed by the ContentsConverter, which does the data format conversion outside of our classes. To display data to the user, the activities access the classes through the database, which can also build them. And to send the classes to the database, another converter is required.

Effectively, because I know that didn't make any sense, the classes just store data, and external converters are required to move it around between formats.

However, I'll have to give this idea some thoughts, because it might be better to internally store the converters to make this easier. We'll see.