Totally rewrite database schema and data model
# Explanation
I realised that the data model I pushed in the last commit was pretty pointless at this stage in the project. My data should be going directly to the database from the parsed XML, and only converted to a Java format once it reaches the UI.

So I had to rewrite everything. At the same time, I thought it sensible to finalise my database columns, and name them appropriately. We also build the CREATE TABLE statement dynamically, which was just a pain in the ass.

# Issues Encountered
I tried to have my program iterate through the classes in DatabaseContract and then iterate again through their methods to dynamically generate tables.

This did not work.

After a few hours, I eventually settled on having lists of keys in each DatabaseContract subclass, and iterating over them. Which means there's a bunch of useless statements in DatabaseHelper which are a result of my poor understanding of Java reflection.

Also I don't really want to understand Java reflection, because it's a pretty niche understanding for such a small part of one project in one language. Not worth the time.