Update counts of children for parent tasks
 - Neatly update all children counts except for 'available'
 - Correctly get the 'next' task to complete in a set
 - Add parameter for 'next' in schema

# Explanation
The database also has a bunch of columns for child counts. These are usually displayed throughout the app, and, well, it doesn't hurt to run a few queries to update them anyway. Not much processing power destroyed.

For each of these, we just run a query to select children of the parent which have the correct parameters (for example, a `completed` date), and then we count the number of rows returned. Easy.

We also need to update the `next` task in a parent, which describes the task that's up for completion next. We can do that by ordering child rows by rank, and picking the first one without a `completed` date.

Also, this is the first commit to be auto-pushed to the scotsict forums! Hi guys!

# Issues Encountered
Firstly, I accidentally thought that getting each child to update its parent, one-by-one, would be the most efficient solution when I have SQL queries. That wasted maybe 15 minutes. Damn.

After that, it was mostly hassle-free (so just hoff), until I finally found that the overdue counts weren't updating properly. I solved this by removing the quotes from the equality test. Not sure if that's actually what did it, because it's late and I can't remember everything.

Peace :)