Fix bug where Omni change their own servers
 - Add a new method to AccountHelper for changing account data
 - Properly nullify the task listener if we can't get files to download
 - If Omni change your data location, update and resync

# Explanation
From time to time, Omni will change the server where your account data is located. For example, I was on `sync11.omnigroup.com` before, and now I'm on `sync7.omnigroup.com`.

We already had code from the login method to handle this, so I copied and pasted it into the file downloader task (it was too small to warrant a new function), and handled the response appropriately. In this case, we just add the new server location to your account data and trigger a resync, which will use the new server location.