Display items from database in basic list
 - Add fragment for a list
 - Make database contract public so it can be accessed by the list creator
 - Create an asynchronous data loader for the database cursor
 - Update tasks on import ID conflict
 - Borrow CursorRecyclerAdapter from Arnaud Frugier (thanks!)
 - Heaps more, this took ages

# Explanation
This was an effort to build. Last one for the holidays though, got to roughly where I wanted to be with this. I'd say at a best guess I'm somewhere between 1/4 and 1/3 done with this, which is really really good!

Anyway, it turns out when you want to display things in a list on Android, it's a real hassle. There's a bunch of old, broken libraries and confusing best practices and I'm still not sure if I might need to replace my RecyclerView with a ListView. We'll see.

To show your database items in a list, you need to:
 1. Open and attach a fragment containing your list (but not a ListFragment, because we're using a RecyclerView) in your Activity
 2. In that fragment, asynchronously load a Cursor with your query
 3. Load a database data-to-view converter (known as an adapter) while you wait
 4. Replace the null Cursor you passed to your Adapter with the one you received from your loader
 5. Let the RecyclerView work its logic through your Cursor, and create/update rows as necessary

This actually takes about 6 hours in real life. I think anyone who's had to work with OS X development would be able to attest to the fact that making a list is actually just really really confusing.

# Issues Encountered
The real problem I had was that I was trying to work with a very specific case that wasn't documented well as a result. See, the RecyclerView was only added about a year and a half ago, and people rarely use it with a database. However, it's the best practice and stops your lists from jamming up when you scroll, so I eventually decided that I had to do it this way (the hard way, as usual).

So most of this was just trying to figure out what the best practice was, and then how to implement it. But eventually, I made it. And I know a lot more about Android development as a result.