Reorganise, Rearrange, Reformat
I've decided I'm never going to be able to handle the mess that some of these classes feel like, but I decided over the past couple hours to fix a few of them a little bit. It's just nicer to be writing code that's properly formatted and is easy to use.

So in this spirit, I:
 - Renamed a bunch of classes
 - Enabled automatic reformatting and rearranging
 - Set up rules for reformatting
 - Split a few classes into easier subclasses
 - Used one logical task per subroutine (at least, in RecursiveColumnUpdater)

Also, these commits should now be automatically GPG-signed on GitHub. Provided the kinks are ironed out (which they... should be?)