Add `blocked`, `hasChildren` and `nextTask` to tree updater
 - Track the next available task in each parent
 - Track `blocked` based on this
 - Set `blocked` on all items in dropped projects
 - Track `blockedByDefer` based on deferrals
 - Track `hasChildren` based on existing child counts
 - Make the database tree loop more efficient by using better SQL statements for cascading values
 - Fix `MainActivity` not noticing that `LoginActivity` has finished
 - General efficiency stuff in `RecursiveColumnUpdater`
 - Recursively fill `projectID`, a new column

A task is 'blocked' in OmniFocus if it can't be worked on right now, because it requires something before it to be completed, it has been dropped, or it's been deferred into the future. This commit tracks blocking through the OmniFocus database and updates columns appropriately.

It's harder than it looks. Lotta algorithm thinking involved. Luckily, I'm pretty hopeful that the algorithm stuff will die down for a bit, at least until I have to worry about editing and syncing again. Once I get onto the UI this gets a lot easier and a lot more fun. I hope.