Recursively update contexts
The next chapter in the Great Recursive Update Saga is updating contexts. A lot of this was pretty much the same stuff (and in some cases quicker/easier) as last time.

Speeds are also looking pretty good—the whole process takes about 7.5 seconds, which is good for a recursive tree traversal like this.

This means the last thing I need to figure out is how to count the number of available items. That is, anything which isn't blocked, blockedByDefer, or on hold. Problem is, we're setting 'on hold' in the context updater, which requires the first loop to have run for its own counter. So we'll probably have to run another loop (contexts can be done in the existing one, probably) to fix it up.