Complete recursive updater
 - Add data types to SQLite database
 - Fix `blocked` not filling down properly by separating it into a
 new loop
 - Count 'available' tasks (not blocked, deferred, on-hold)
 - Speed up project counting by separating it, also enabling
 'available' counting
 - Remove a useless test for the size of a ContentValues that would
 always return true
 - Re-implement the boolean tracking of whether an item has children
 or not
 - Correctly fill down due dates, defer dates & flags
 - Correctly track project status and type as the loop fills down

And that's a wrap! I've checked out this update on my phone and everything (almost) is being appropriately tracked and counted, at least to the point that would satisfy any OmniFocus user who's game enough to use it on Android.

While updating and fixing this subroutine was a slog and a half, it's rewarding because it means I can finally get working on the UI and the important-yet-possible-easier bits of the application. Which is great for everyone!