Initialise data model
 - Create data model for all data types except perspectives
 - Update Android Studio to 2.1, and a few SDK updates

I've decided that the way I'm going to sort everything out is to read a list of items from the database into their own classes (like an array), and display that list to the user through a RecyclerView (like a ListView).

To do that, I needed to build a proper model of all the classes which might be used. This model is totally incomplete—there's lots of data types which might change as I discover new things which are more helpful for displaying data, and there's going to be a lot of helper methods added. But this way, I can have an array of contexts / tasks / folders / whatever to display to the user.