Create helper class for data model
To actually use the data model, we need classes that can convert database entries into instances of the data model's classes

In this commit, our class is able to, given a parent ID, return lists of children of different types of items (contexts, projects, folders) which can be used to populate various views.

Keep in mind that this is totally incomplete for the moment—I need to add classes like an inbox getter and the like—but it'll do for building the views that we need.

 - Add `DataModelHelper`
 - Add cursor-to-data-model methods
 - Add get-children-of methods
 - Add column lists to the database contract
 - Add `getLong()` and `getDate()` methods to the database helper class
 - Change the type of `projectID` in the Task class