Create basic list view
 - Add list fragment
 - Add fragments for individual items
 - Add an adapter for a list of tasks

Wow!

We can check off the backend stuff as Done For Now, and move onto the more exciting frontend. With that in mind, I finally got over the ListView/RecyclerView hurdle and have built a list of items!

At the moment it just displays a basic list of tasks that have no parent (so top-level projects and things that haven't been deleted yet), but soon it'll do all the rest of stuff.