Add a navigation drawer
That was like 200x more difficult than it should've been. I spent a number of hours trying to fix various layout and UI bugs, glitches and the like. It's really difficult but almost entirely necessary to support older versions of Android, which makes the job a lot more difficult—you can't just type in the simple classes and have them work.

Plus, Google are never really clear with what exactly is the best practice to use. You have to do about 15-30 mins of research before commencing anything, which hurts a little. But it's well worth it.

At the end of this process I've built a nav drawer. It's something I can interact with in the app. It's really satisfying. And soon I'll be able to press some of these buttons and have various views appear. And at that point it'll be looking a lot closer to complete.

There actually isn't all that much left to go on this project, really. I can see it being done by August if I put the time in. That's a good feeling.

No changelog because I need to sleep.