Change theme and menu item color in nav drawer
If you open up Google Inbox on your phone, you'll notice the icons in the side menu are coloured, and when you press on them, their highlight colours reflect that. This is actually not something which Android likes to advertise it can do and it's not a default.

To work with this, I had to create a 'ColorStateList', which is a dictionary of different item states (focused, selected, checked, etc.) and corresponding colours. Once someone selects an item in the sidebar, all of the colours change to that item's colur, but only this item is selected so you only see it change.

For good measure, we also change the app's theme—that means its accent colours and the rest. Next stop is to change the toolbar at the top of the app, and we can apparently animate it! Exciting!