Theme up the navigation
 - Colour app bar and status bar, and transition them according to
 the currently selected item
 - Lighten the background highlight on the currently selected item

Took a lot longer than you might think, because it's really confusing
 to deal with status bars and the rest. I first experimented with
 setting a padding on the toolbar at the top of the screen, but this
 messed things up when I turned on immersive mode/expanded desktop.

The way to solve it, eventually, was to transition the colours with
the toolbar. Apparently Android can take care of the status bar
colour just fine without you having to set a darker variant. The
other thing to remember was that I had to call `drawerLayout
.invalidate()` _after_ I set the status bar colour, to refresh the
view.