Implement perspective switching
 - Add a flexible `getContexts()`, `getFolders()` and `getTasks()`
 method for retrieving entries from the database given specified
 selections and arguments
 - Re-implement current data model getters using this function
 - Add getters for tasks that're currently in the inbox, are top
 level projects, or flagged
 - Rename `TaskAdapter` to `ListAdapter`, and adapt it to any item
 (not just tasks) (semi-done)
 - Vary the items picked up by the ListFragment depending on
 arguments passed by the calling function (will need to change again
 for nested lists)
 - Set up toolbar and navbar in `MainActivity` in their own subroutines
 - Switch between fragments on menu item selection
 - Show a progress spinner in the interim
 - Correctly set `colorAccent` for the spinner colour

The navigation menu from previous updates now does more than just change the colour, it changes between different OmniFocus perspectives, as represented by android Fragments (like a cross between a View and an Activity). Effectively, when you click on an item in the sidebar, the current view will fade out, a progress spinner will fade in, and then fade out again when the new Fragment is ready to show up.

One of the bigger problems with this was working out the UX that Google uses in their Inbox app, which switches fragments and uses a coloured toolbar in a similar way. By default when switching fragments, you switch immediately on press, which is a jarring and non-animated transition. It's also really laggy.

Instead, I realised that Google fade out on click, and fade in once the drawer has closed (while also displaying a spinner as it loads the content). This solution seemed to work fairly well for me, and it creates a smooth and comfortable transition between views.

The next steps are a little more confusing: I need to be able to implement composite lists of Tasks, Contexts and Folders, and display them in each context—right now it breaks when we select 'Contexts' and doesn't show top-level Folders. This will be a challenge, particularly because they have to use the same adapter for the whole list, but I'm sure I'll figure it out (more like, I _have_ to figure it out).