Allow different types of view in list
 - Rename 'Entry' to 'Entries' in the database contract to avoid
 conflict
 - Add some more SQL shorthands to the data model helper
 - Get a readable database instead of a writable one when all we need
  to do is read from it (in the data model)
 - Generalise all returns from the data model to `List<Entry>`
 - Add the 'No Context' context to the list of top level contexts
 - Add a `getTopLevelFolders()` method to the data model
 - Add folders to the `getTopLevelProjects()` method
 - Update folder child counts in the RecursiveColumnUpdater
 - Update the available children count in the context updater
 - Implement `Comparable` in `Entry`, which compares by rank
 - Extract ViewHolders for Tasks, Folders and Contexts
 - Implement multiple ViewTypes and corresponding logic in the List
 Adapter
 - Add Due Soon and Overdue backgrounds to the Task list item
 - Oh man

So I know this requires a huge explanation, but it's almost 1am and I need to get some sleep. Plus, I don't think I can write and make much sense at the same time right now. But let's give it a try because I wanna be done with this for the weekend.

Because OmniFocus mixes its lists up, such that we can often see a Folder and a Task in the same list, we also need to mirror this functionality on Android. It's crucial that we can see everything the same way.

On Android, all we have to do is implement different View Types in the RecyclerView, which helps us instantiate new ViewHolders as we need them, with different types depending on what we want to display. In the ListAdapter, we look ahead (using the object's position) at what type it'll be, and generate a ViewHolder based on that, using one of the lovely views and classes we have.

In future, this will need to adapt to whether the task has children or is a project or whatnot, too. Keep an eye out.

Once this is done, it's not too difficult to just take a list of entries, split them up by type, and add each type to the ViewHolder it wants dynamically. Which is what we do.

The problems struck, I think, when I wanted to implement new things, like Folders. The chief one I dealt with just then was getting the folders to display counts of their children. This is tricky because of the rules around folders and projects:

 1. A folder can hold other folders
 2. A folder can hold projects
 3. Projects cannot hold other projects

So the way to solve this while maintaining some efficiency was to have each project update its parent's (which must be a folder or nothing) child counts with its own. But we do this with a simple dictionary of folder IDs and child counts. Then, we iterate through each item in that dictionary and do the same until there are none left, adding items along the way. Finally, we take each ID in that list and update its child counts in the database.

This saves energy because we don't have to count any new items, just push old counts up the stack. There's probably a quicker way to do it. I don't care right now; I need to sleep. Have a nice weekend.