Add empty state
When a list is empty, it's important to let the user know the app hasn't broken and is just showing them an empty view. For that reason, we use empty states—placeholder views in the background of the list, which are activated if there aren't any list items.

The new code checks the numbers of items in the RecyclerView on creation, and if that number is 0, then it displays a pretty simple logo and message to let the user know. The UI on this section isn't final, but it likely won't change before the submission date.