Add rows for projects and nested tasks
In OmniFocus, projects and nested tasks (i.e. tasks with child tasks) should display differently, because they're containers of tasks instead of tasks themselves, and display in different contexts (this time, I use 'context' in the traditional English sense).

A project displays exactly like a folder, with a 'remaining' count + an 'overdue' and 'due soon' count. The code is pretty much cut-and-paste and isn't going to change for the moment, but I kept it separate because I think they'll need to be separate things in the future.

A nested task displays like a hybrid of the two, with an 'x remaining' counter but also a due date. For this reason, I cut off the overdue/due soon counters because they'd fill the view. When I re-implement them, I'll likely structure it like the iOS OmniFocus app, which uses dots.

Finally, I also moved the ViewHolder binding classes (i.e. setting up the fields with the given task/context/folder) into the ViewHolders themselves, which cleans up `ListAdapter`.