Discolour unavailable tasks
In normal OmniFocus, the user wants to be able to skim-read a list and understand as much as they can, because they're a power user and love that sort of stuff. This is why a task that isn't 'available' (i.e. blocked or completed) is greyed out, which is what I just did.

Now onto traversing the big list of tasks/projects/contexts/folders!