Enable full tree traversal
 - Add methods to `DataModelHelper` for getting the appropriate
 children of an entry
 - Allow 'No Context' to be selected and displayed
 - Add a new initialiser to `ListFragment` that appropriately gets
 the data for an individual item
 - Reenable `OnListFragmentInteractionListener` and implement
 - Enable touch ripples in the RecyclerView, and convert margins to
 padding to suit
 - Use `FragmentTransaction.replace()` instead of removing, to remove
  all fragments
 - Change the default theme in `MainActivity` to the red theme (given
  that most users will start on the Forecast perspective, and not
  need to see the transition from purple.

The basic viewing app is almost complete! I've come a _long_ way today in putting this together, and now the user can entirely traverse the project and context trees—starting from the base elements and reaching whatever they need to (provided it isn't completed, in which case it's hidden).

The only real step I need to take now is to build the detail views in the list, and I'll have a rudimentary but fully-functional viewer app. Then I'll have to worry about editing, then syncing back up, so what I've decided is that from here I'll try to skip the little details, like animations between screens and fancy colouring, until the app is complete. I don't need those UI flourishes for the assessment, but I will need them for market, and I'm lucky that those are two separate stages in this process.

So I'm going to try and do as much as I can, as quickly as I can, and then move onto the filling in the fancy details later.