Create basic Detail View
# Important Update
I've decided that for the purposes of this assignment, I'm not going to complete a fully syncing OmniFocus application, just a viewer. The whole thing is going to take up far too much of my time, and I won't be able to get that done by August without failing the rest of my subjects.

From now until the due date, I'm going to complete the detail viewer, and add a notification system. I'll also need to implement an automatic syncing routine with that, and there'll probably be other stuff on the way. Once that's done, I'll add some simple finishing touches, and complete the rest of the project once school's out.

It's been really fun so far and I'm looking forward to where this is going. It doesn't just look like a sample project anymore, but instead it really feels like an app I'd use on a semi-regular basis. Once I get some more stuff into it, hopefully that'll only get stronger.

# Changelist
 - Make all data model items Parcelable
 - Appropriately handle null dates in the parceling process with new
 methods
 - Rename the various fields in the `ViewHolder`s to `entry`
 - Add a `DetailActivity` for displaying task/context/project details
 - Change the colour of `DetailActivity`'s app bar depending on where
  it was launched from (to maintain visual continuity).
 - Replace `EditText` with `TextInputEditText` for better support
 - Perform null checks for UI elements
 - Get the toolbar and drawer layout in `MainActivity` if they're null
 - Colour the `TextInputLayout` in the toolbar of `DetailActivity`
 appropriately

# Explanation
This change took significantly longer than it needed to. The main reason for this is the way that Android messes around with the status bar and app layouts. Suffice to say I was not happy.

In a `DrawerLayout`, Android appropriately determines that you want the navigation drawer to display underneath the status bar, and extends the UI below it appropriately. This extension includes the toolbar.

Inexplicably, it doesn't do this for any other type of layout. So I spent a few hours trying to figure out what was actually going on with my UI, which was difficult because I had no idea what to search for and my internet was slow because I'd inadvertently downloaded a botnet. Or maybe it was just slow because it slows down sometimes—I'm really not sure.

The solution, eventually, was to set the background colour of the view to be the same as the toolbar, so it looks like the toolbar extends below the status bar. Then we set the background of the rest of the view (underneath the toolbar) to be the normal android background (which I think is white). That way, we get the same effect as before, but without having to define a new status bar colour (because it doesn't tint your colours automatically).