Implemented decryption
For this update, I wrote an OmniFocus file decrypter based on the
mdoel that OmniGroup wrote in Python. It took a number of days and
was one of the heftier parts of the project, particularly considering
 I had little idea on how to use Java in the ways I wanted to.

But, eventually, we got there. This update adds a number of minor UI
changes (including the removal of the `Nearby` menu item, since it's
no implemented), and the addition of this routine. It's not perfect
and will probably throw errors and have a number of security
holes—and let's just hope nobody tries to use it with an older
version of OmniFocus—but it works for my database on various
phones/tablets so that should do for now.