Move sync to background, rewrite parser
The synchronisation loop has been moved out of its original class and into a `SyncAdapter`, which is a more appropriate Android loop and allows us to correctly handle synchronisation (instead of manually / called whenever I want it called). This was basically the last step of this project, once I add the stuff that gets it calling the sync loop automatically, and displaying a spinner on MainActivity whenever the user is syncing.

Also, when doing that, I found that the XML parser was broken. Instead of trying to save XMLPullParser, which is honestly the most horrifying tool I've had to use during this project, I switched to the w3c DOM classes, which work a lot better—although required a lot of finnicky rewriting. On top of that, Omni changed a lot of the sync types, which made this a lot harder. But lo, it was finished too.

Now to finish up syncing...