SDD hand-in done
 - Close leaking SQL cursors
 - Test for accounts before loading swipeLayout
 - Fix bug where recyclerView would stay invisible after load (took me AGES)
 - Update spinner if actively syncing
 - Send DetailActivity intents back to DetailActivity
 - Correctly send LoginActivity result to MainActivity
 - Correctly receive results and start sync
 - Register periodic syncs and turn on automatic syncs
 - Increase minimum SDK version to prevent untested code from working on old platforms

This should be the final version of the application before hand-in on 19/8. I've tested it on my phone and tablet running 5.1, and the new sync loop totally works. And it also lets the user know when we're syncing! Plus there's enough internal help to fill the requirements as far as I'm willing to go.

This is exciting—as soon as I figure out how to export an APK, I'll upload it. Wouldn't hurt to test this.