Correctly pass Context to AccountManagerHelper
 - Update SDK and compilation sources to API 24 (Android 7.0 Nougat)
 - Pass Context to `AccountManagerHelper`
 - Reorder the arguments passed to `DownloadFileTask`

In the previous version, `AccountManagerHelper` relied on a context saved in `MainActivity` to run. This broke when a background sync was triggered, because the app was no longer open and thus the context was inaccessible. In this version, the context is retrieved from the sync adapter and sent to the account manager. In other places where the account manager was used, the context is now correctly passed to it natively. This is the more sensible way of running the code.

I have left the `context` field in `MainActivity` because certain classes still rely on it. However, none of these classes need to run in the background and should always be able to access MainActivity's context, so I don't see any harms arising. This will obviously be fixed before a full release.