Speed up decryption by saving key
In the decryption routine we derive an AES-wrap key from our
passphrase, which slowed down the syncing routine by at least a
minute. Since this is simply a process of converting the plaintext
passphrase into a more-easily-used AES key, it's safe to save into a
file and read later. If that doesn't work, we just re-derive it.