Better integrate data model
The data model is now more heavily integrated—instead of using
references to IDs, we now use strict references to each item when
populating lists. It saves a bit of time and makes sure that we're
getting the right objects, plus gives us room for better features in
the future.