Finish rewriting RecursiveColumnUpdater
I basically just added and refined some methods from
RecursiveColumnUpdater, and added some stuff for updating and
counting contexts. It's a bit late, but I shaved a second off the
methods I added. The whole thing's pushed about 2 seconds on top, but
it's fine because it works better. All good.