Implement CardView list
- Create HeaderViewHolder to reflect widening differences between
headers and list items
  - Remove headerRow field
  - Remove getNameTextStyle() method
- Split FragmentInteractionListener into a listener to open a list,
and a listener to open a detail view
- Move all list items to CardView, so now it looks like Google Inbox
- Merge in Google's DividerItemDecoration, but replace with
OnDrawOver (so it draws over CardView shadows)

This was a nightmare, so just to be clear: Google's inbox thing
relies on header rows which aren't CardViews, and item rows that are.
Each item has a divider which is very small, and draws _over_ the
RecyclerView (otherwise the shadows from the cards draw on top of it).
Inside the CardView (which is just a FrameLayout) is the normal
RelativeLayout. The background is a nice grey, and only the CardViews
have a white set to their background (with a ripple foreground).