Better handle the back stack
This now works like Google inbox. We have a specific 'home' screen
defined (which is the checked drawer item, or the first in the
perspective list-- 99% of the time this is the Forecast). When we're
in this screen, pressing back closes the app. In any other
perspective, it returns to this perspective. This is more in line
with what the user expects to do, and more in line with what appear
to be best practices.